+++ src/fonts/font-definitions.ts
@@ -17,7 +17,21 @@ export type BuiltInFontId =
   | "rochester"
   | "sacramento"
   | "satisfy"
-  | "yellowtail";
+  | "yellowtail"
+  | "edu-nsw-act-foundation"
+  | "edu-vic-wa-nt-hand"
+  | "league-script"
+  | "playwrite-be-vlg-guides"
+  | "playwrite-ca"
+  | "playwrite-de-grund-guides"
+  | "playwrite-id"
+  | "playwrite-mx"
+  | "playwrite-nz-basic"
+  | "playwrite-nz-basic-guides"
+  | "playwrite-nz-guides"
+  | "playwrite-pt"
+  | "playwrite-ro-guides"
+  | "playwrite-us-trad";
 
 export interface BuiltInFontDefinition {
   readonly id: BuiltInFontId;
@@ -38,6 +52,90 @@ export const BUILT_IN_FONTS: readonly BuiltInFontDefinition[] = [
     "Architects Daughter",
     "architects-daughter/ArchitectsDaughter-Regular.ttf",
     "Kimberly Geswein",
+  ),
+  font(
+    "edu-nsw-act-foundation",
+    "Edu NSW ACT Foundation",
+    "edu-nsw-act-foundation/EduNSWACTFoundation-Regular.ttf",
+    "The AU School Handwriting Fonts Project Authors",
+  ),
+  font(
+    "edu-vic-wa-nt-hand",
+    "Edu VIC WA NT Hand",
+    "edu-vic-wa-nt-hand/EduVICWANTHand-Regular.ttf",
+    "The VIC WA NT School Hand Australia Project Authors",
+  ),
+  font(
+    "league-script",
+    "League Script",
+    "league-script/LeagueScript-Regular.ttf",
+    "Haley Fiege",
+  ),
+  font(
+    "playwrite-be-vlg-guides",
+    "Playwrite BE VLG Guides",
+    "playwrite-be-vlg-guides/PlaywriteBEVLGGuides-Regular.ttf",
+    "The Playwrite Project Authors",
+  ),
+  font(
+    "playwrite-ca",
+    "Playwrite CA ExtraLight",
+    "playwrite-ca/PlaywriteCA-ExtraLight.ttf",
+    "The Playwrite Project Authors",
+  ),
+  font(
+    "playwrite-de-grund-guides",
+    "Playwrite DE Grund Guides",
+    "playwrite-de-grund-guides/PlaywriteDEGrundGuides-Regular.ttf",
+    "The Playwrite Project Authors",
+  ),
+  font(
+    "playwrite-id",
+    "Playwrite ID ExtraLight",
+    "playwrite-id/PlaywriteID-ExtraLight.ttf",
+    "The Playwrite Project Authors",
+  ),
+  font(
+    "playwrite-mx",
+    "Playwrite MX ExtraLight",
+    "playwrite-mx/PlaywriteMX-ExtraLight.ttf",
+    "The Playwrite Project Authors",
+  ),
+  font(
+    "playwrite-nz-basic",
+    "Playwrite NZ Basic ExtraLight",
+    "playwrite-nz-basic/PlaywriteNZBasic-ExtraLight.ttf",
+    "The Playwrite Project Authors",
+  ),
+  font(
+    "playwrite-nz-basic-guides",
+    "Playwrite NZ Basic Guides",
+    "playwrite-nz-basic-guides/PlaywriteNZBasicGuides-Regular.ttf",
+    "The Playwrite Project Authors",
+  ),
+  font(
+    "playwrite-nz-guides",
+    "Playwrite NZ Guides",
+    "playwrite-nz-guides/PlaywriteNZGuides-Regular.ttf",
+    "The Playwrite Project Authors",
+  ),
+  font(
+    "playwrite-pt",
+    "Playwrite PT ExtraLight",
+    "playwrite-pt/PlaywritePT-ExtraLight.ttf",
+    "The Playwrite Project Authors",
+  ),
+  font(
+    "playwrite-ro-guides",
+    "Playwrite RO Guides",
+    "playwrite-ro-guides/PlaywriteROGuides-Regular.ttf",
+    "The Playwrite Project Authors",
+  ),
+  font(
+    "playwrite-us-trad",
+    "Playwrite US Trad ExtraLight",
+    "playwrite-us-trad/PlaywriteUSTrad-ExtraLight.ttf",
+    "The Playwrite Project Authors",
   ),
   font(
     "allura",